+++ com.unity.renderstreaming/Editor/SerializedPropertyExtension.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2e226116f6c4a4b40a27c93778cd162c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
